refactor(jobs): remove unused import and optimize column definitions
Removed an unused import (useHTMLToPlainText) from the columns.tsx file and optimized
the ColumnDef structure for Jobs by eliminating unnecessary React imports and a cellcomponent that were previously used but no longer required. Additionally, in the select.tsx
component, the focus ring color was set to transparent by default to maintain a clean UI
aesthetic when the select input is focused.